refactor(risk): share the per-trade risk cap with the backtest engine
The live per-trade risk cap (commit 1767053) did not apply to the offline
backtest: BacktestEngine / PositionSizing.sizeEntry used `decision.quantity ?:
riskSized`, so a recorded AI decision backtested at its raw quantity. Replaying
one recorded BTC decision lost 34% of a $10k book on a single trade — a
position the live cap would have shrunk.

Move `resolveRiskCappedQuantity` from application/usecase into domain.services
(the pure sizing rule now has ONE home, shared by the live path and the
backtest) and route sizeEntry through it. Rule strategies (no quantity, no
risk_usd) are unaffected — the function reduces to the identical
`equity*pct/|fill-stop|` formula with the same 8dp HALF_UP rounding. Bump the
backtest defaultConfig riskPerTradePct 0.01 -> 0.02 to match the live default
(RiskGateProperties.maxRiskPerTradePct).

Now replaying the SAME decision over the SAME candles caps the loss at -2.02%
(the 2% budget) instead of -33.75% — sign preserved, magnitude capped, so the
backtest reflects what the bot would actually trade. Test moved with the
function (10 cases) + a new clamp test in PositionSizingTest; BacktestEngineTest
disables the cap (riskPerTradePct=0) to keep its hand-computed curve. Full suite green.
