+++ src/main/kotlin/ru/driics/aitrade/application/usecase/ExecuteAiDecisionsUseCase.kt
@@ -20,6 +20,7 @@ import ru.driics.aitrade.domain.ports.TradeJournalPort
 import ru.driics.aitrade.domain.ports.TradingPort
 import ru.driics.aitrade.domain.services.IdGenerator
 import ru.driics.aitrade.domain.services.OrderSizingPolicy
+import ru.driics.aitrade.domain.services.resolveRiskCappedQuantity
 import ru.driics.aitrade.domain.types.InstrumentId
 import ru.driics.aitrade.domain.types.InstrumentResolver
 import ru.driics.aitrade.domain.types.OrderSide
@@ -562,46 +563,4 @@ internal fun selectEntryPrice(
     if (wsFresh == null) return restPrice
     if (priceDeltaBps(wsFresh, restPrice) > maxDeltaBps) return restPrice
     return wsFresh
-}
-
-/**
- * Per-trade position sizing with a hard risk ceiling — pure + pinned because it picks the coin quantity
- * that sizes a real order. Makes `risk_usd` authoritative: the budget is `min(requestedRiskUsd,
- * maxRiskPerTradePct * equityUsd)`, the quantity is `budget / |entry - stop|`, and the model's own
- * [modelQuantity] is advisory — honoured only when it is MORE conservative (smaller) than the budget.
- *
- * Returns 0 (the caller skips) whenever the cap cannot be enforced — no stop, zero/negative equity, or a
- * non-positive budget — so an inflated model quantity is never traded uncapped. [maxRiskPerTradePct] <= 0
- * disables the cap and restores legacy behaviour (model quantity wins; size from [requestedRiskUsd] else).
- */
-internal fun resolveRiskCappedQuantity(
-    modelQuantity: BigDecimal?,
-    requestedRiskUsd: BigDecimal?,
-    equityUsd: BigDecimal,
-    maxRiskPerTradePct: BigDecimal,
-    entryPrice: BigDecimal,
-    stopLoss: BigDecimal?,
-): BigDecimal {
-    fun quantityForRisk(riskUsd: BigDecimal?): BigDecimal {
-        if (riskUsd == null || riskUsd.signum() <= 0 || stopLoss == null || stopLoss.signum() <= 0) {
-            return BigDecimal.ZERO
-        }
-        val riskPerUnit = (entryPrice - stopLoss).abs()
-        return if (riskPerUnit.signum() > 0) riskUsd.divide(riskPerUnit, 8, RoundingMode.HALF_UP)
-        else BigDecimal.ZERO
-    }
-
-    // Cap disabled -> legacy: the model's quantity is authoritative, risk-sizing only as a fallback.
-    if (maxRiskPerTradePct.signum() <= 0) {
-        return modelQuantity ?: quantityForRisk(requestedRiskUsd)
-    }
-
-    val maxRiskUsd = (equityUsd * maxRiskPerTradePct).coerceAtLeast(BigDecimal.ZERO)
-    // Budget is the model's requested risk capped at the equity-based ceiling; the ceiling alone when the
-    // model gave no (positive) risk_usd.
-    val budget = requestedRiskUsd?.takeIf { it.signum() > 0 }?.min(maxRiskUsd) ?: maxRiskUsd
-    val cappedQty = quantityForRisk(budget)
-    if (cappedQty.signum() <= 0) return BigDecimal.ZERO
-    // Model quantity is advisory: keep it only when it risks LESS than the budget.
-    return modelQuantity?.min(cappedQty) ?: cappedQty
 }
